backend/pyproject.toml
[project]
name = "fxsociety-backend"
version = "0.1.0"
description = "FastAPI backend for fxsociety trading marketplace"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "sqlalchemy>=2.0.0",
    "pydantic[email]>=2.0.0",
    "pydantic-settings>=2.0.0",
    "python-multipart>=0.0.9",
    "python-jose[cryptography]>=3.5.0",
    "passlib[bcrypt]>=1.7.4",
    "argon2-cffi>=25.1.0",
]

[project.optional-dependencies]
dev = [
    "httpx>=0.27.0",
]

[tool.uv]
dev-dependencies = [
    "httpx>=0.27.0",
]
